compilation errors fixed in Windows

+++ bg2-math/include/bg2math/vector.hpp
@@ -84,10 +84,10 @@ namespace bg2math {
         return v1[0] * v2[0] + v1[1] * v2[1];
     }
     template <class T> inline bool isnan(const vec2<T> & v) {
-        return std::isnan(v[0]) || std::isnan(v[1]);
+        return std::isnan(static_cast<double>(v[0])) || std::isnan(static_cast<double>(v[1]));
     }
     template <class T> inline bool isinf(const vec2<T> & v) {
-        return std::isinf(v[0]) || std::isinf(v[1]);
+        return std::isinf(static_cast<double>(v[0])) || std::isinf(static_cast<double>(v[1]));
     }
     template <class T> inline bool iszero(const vec2<T> & v) {
         return v[0] == 0 && v[1] == 0;
@@ -195,10 +195,10 @@ namespace bg2math {
         return v1[0] * v2[0] + v1[1] * v2[1] + v1[2] * v2[2];
     }
     template <class T> inline bool isnan(const vec3<T> & v) {
-        return std::isnan(v[0]) || std::isnan(v[1]) || std::isnan(v[2]);
+        return std::isnan(static_cast<double>(v[0])) || std::isnan(static_cast<double>(v[1])) || std::isnan(static_cast<double>(v[2]));
     }
     template <class T> inline bool isinf(const vec3<T> & v) {
-        return std::isinf(v[0]) || std::isinf(v[1]) || std::isinf(v[2]);
+        return std::isinf(static_cast<double>(v[0])) || std::isinf(static_cast<double>(v[1])) || std::isinf(static_cast<double>(v[2]));
     }
     template <class T> inline bool iszero(const vec3<T> & v) {
         return v[0] == 0 && v[1] == 0 && v[2] == 0;
@@ -314,10 +314,10 @@ namespace bg2math {
         return v1[0] * v2[0] + v1[1] * v2[1] + v1[2] * v2[2] + v1[3] * v2[3];
     }
     template <class T> inline bool isnan(const vec4<T> & v) {
-        return std::isnan(v[0]) || std::isnan(v[1]) || std::isnan(v[2]) || std::isnan(v[3]);
+        return std::isnan(static_cast<double>(v[0])) || std::isnan(static_cast<double>(v[1])) || std::isnan(static_cast<double>(v[2])) || std::isnan(static_cast<double>(v[3]));
     }
     template <class T> inline bool isinf(const vec4<T> & v) {
-        return std::isinf(v[0]) || std::isinf(v[1]) || std::isinf(v[2]) || std::isinf(v[3]);
+        return std::isinf(static_cast<double>(v[0])) || std::isinf(static_cast<double>(v[1])) || std::isinf(static_cast<double>(v[2])) || std::isinf(static_cast<double>(v[3]));
     }
     template <class T> inline bool iszero(const vec4<T> & v) {
         return v[0] == 0 && v[1] == 0 && v[2] == 0 && v[3] == 0;
